Recomputed System Tests for correctness and reorganized
The system tests didn't have the local ribs, which was a problem for actually checking if the system is working with pytest. The current configuration of the simulation is not set for actually doing long runs (its setup for correctness, and its slow). 

Reorganized the way the system tests to allow parsing into a directory structure for overleaf easier. That will come in a later commit.