Fixed a bug for k=0
Realized that adding the reporting AS in the path does not yield intended behavior for k=0. This was realized in test cases by_bgp. For example in graph011 an AS that we once connected with an alternative path became disconnected, just because itself was on the avoid list because it reported receiving the hijack.  Really, it should continue using its best alternative path and not disconnect itself just because it has received the hijack.