Added an assert statement to verify Disconnected Avoid List case
If an avoid list ASN tracebacks to a DISCONNECT outcome, it should be due to a blackhole. If that's not the case, then it's error. An assert statement has been added to check this case.

Graph class now prints progress as it has done before modifying it for collecting reports.