Add fix for tuple network address (Justin helped find this)
With Justin I found a typo that caused the blackhole announcement prefix was created as a tuple. That's resolved in this commit.

Also updated the main script to be deterministic, and print out more useful timing information when running. 

Removed debugging print statements.

Next need to do some performance testing.