Created a seemingly working version of the AutoImmune Attack
Thanks to the help of Justin amkinga change in the base bgp_simulator_pkg I was able to get access to the engine; which gave me access exactly to what I needed. I made a test config 133 to see if it works. Seems to be creating the announcement correctly. This test will need to be updated to properly test this attack.