Updated all tests outcomes and verified them. Added new test for new atk
The test outcomes had to be updated because of a bug found in the mvdp module. The bug was an incorrect mapping of the artificial source and real_asns. 

 I verified all the tests visually, and everything passes. The change probably came from an update in bgp_simulator.

Seem to have implemented the autoimmune attack with just one attacker for now. Need to work on multiple attackers case. 

Added report graph generation to the system tests. It will output all the target_asn graphs to the current directory. This is the current solution until I can find a way of just creating an image for each hijack prefix, and putting them inside the corresponding test folder.