Created the overlay/relay feature and added system tests for it
The overlay/relay implementation implemented assumes there's at most 1 CDN in the relay set. Additional code needs to be added for support more than 1 CDN, but a lot of the leg work for it is already there. 

Also added two system tests for this, and they appear to work as expected!