Added ability for multiple CDNs and fixed a bug in recomputing outcomes
Now multiple CDNs can be specified and used.
Fixed a bug when computing the outcomes.
It turns out I forgot to consider the outcomes of those not adopting. After some adopting ASes become connected, the outcomes of those connected to it should also be recomputed. So now everything that is disconnected after applying relays to the adopted, has traceback redone for it (to update its outcome).