Potentially solved invalid vertex bug
The error occurred because the reported path had a loop in it.

This occurs because the receive_ann function in ROVSMS reports a subprefix prepending the receiving ASes asn without considering that the AS would just drop this announcement it's on the path.

Now ROVSMS will not report if its already on the path.