Attack Relay Setting Added and Fightback Option
Attack Relay Setting added and tested for V4 subprefix hijack scenario. It has not been tested for AutoImmune Attack yet. Two test cases for Subprefix attack with and without attack relay added (62 and 63 respectively). It shows that after attacking the relays, the effectiveness of the relays is reduced. 

Fight back option added to V4 Subprefix Hijack Scenario. It has not been tested, and may not be a needed function anymore after  discussing the impact of the idea last week. It's function is being kept just in case there's a change in plans. Otherwise, it's just dead code if the setting is not invoked.