Reran benchmark on HPC
The benchmark performance is about the same, except with an increase in memory usage. That raise in memory could be a problem, but it's not completely clear to me now how significant that is. Otherwise, the runtime is the same.

My conclusions from this result is that the benchmark is flawed. The performance problem is not evident from the settings being used.

I believe that a better benchmark simulates exactly a large set of runs.