Fixed a bug in how available ASes are calculated
The check for the blackhole was not done properly. The `_relay_is_available` method returned True almost always, because it did not check the relay_asn announcement, but rather the attack announcement it self, and the if condition check was wrong as well. This means all the relay runs need to be redone agian. This had to happen anyway, due to new requested changes anyway.