Fixed a bug for attack relay setting with direct autoimmune
The bug occurred in this direct autoimmune setting attacking relay because theirs a part where the providers are a target of the indirect attack; which it didn't take into consideration in the direct setting attacking the relays. Instead of attacking the providers of the relays, it's simply a direct attack on the relays.