Made a fix for the removing-blackhole issue
The problem was the odd case I didn't pin down when I find blackholes in ribs without them being labeled blackholes. This was a misinterpretation and a not careful condition of checking if something is a blackhole.

There persists an issue for the case of superprefix+prefix hijack prefix, because of the prefix attack part. That will be resolved in the next commit.