Resolved Bugs Affecting CDN Relay Setting and ROVPPO
## Bug 1
This bug was noticed when looking at results for attack_relay setting. What was noticed that the CDN results were the same as the results as if the relays were not attacked.
The bug was in `_recalculate_outcomes_with_relays`. There was a conditional to reconnect disconnected relays for CDN relays. The condition was not setup properly, so it also reconnected any adopting AS to a relay without needinga path to the relay. This impacts any AS that was actually disconnected due to not having a path to the overlay. This may not have a significant impact on the results without attacked_relay, but will have impact on results for attacked_relay = True.

## Bug 2
Bug was noticed in system test 170
This is a test for ROVPPO where most of the CDN relays should have disconnected  (the original result of this test was incorrect, and has been updated in this commit), because the conditional for reconnected CDNs shouldn't have been activated for this setting; which it accidentally was due to an incorrect conditional for setting available_relays.