Converted Config 150 and Created new Graph Class
The conversion of 150 is not runnable yet, but is converted properly. This can be used as an example to convert 162-172, but becareful. GPT may missing a variable or two.

When you run this, you will need to use the latest version of the BGPy simulator.

Running a test now will result in an error for converting some graphs. That's going to be additional tasks that need to be done for this refactor.